New endpoint, deleted unnecesary code, and unnecesary endpoint
