+++ backend/routes/index.js
@@ -7,9 +7,7 @@ const loginRouter = require("./login.router");
 const filesRouter = require("./files.router");
 const usersRouter = require("./users.router");
 const consolidadoRouter = require("./consolidado.router");
-const informationRouter = require("./information.router");
-
-
+const fuentesRouter = require("./fuentes.router");
 
 
 const routerApi = (app) => {
@@ -17,15 +15,12 @@ const routerApi = (app) => {
 	app.use(`/${properties.get("app.api.structure")}/v1`, router);
 
 	// Routes
-	router.get("/", (request, response) => {
-		response.send({message: "Servidor backend para el proyecto Bodega"});
-	});
 
 	router.use("/login", loginRouter);
 	router.use("/files", filesRouter);
 	router.use("/users", usersRouter);
 	router.use("/consolidado", consolidadoRouter)
-	router.use("/info", informationRouter)
+	router.use("/fuentes", fuentesRouter);
 }
 
 module.exports = routerApi;
