--- frontend/src/Pages/components/MessageCard/styles.css
@@ -1,24 +1,0 @@
-.card-container {
-	width: 100%;
-	padding: 20px;
-
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-
-	border-radius: 10px;
-
-	box-sizing: border-box;
-}
-.error {
-	background-color: var(--error-container-color);
-}
-.all-ok {
-	background-color: var(--allok-container-color);
-}
-.card-container p{
-	width: 100%;
-	text-align: center;
-	color: var(--white);
-}
